+++ app/sitemap.ts
@@ -1,20 +1,17 @@
-import { compareDesc } from 'date-fns';
 import { siteConfig } from '@/config/site';
 import { env } from '@/env.mjs';
+import { getAllPosts } from '@/src/lib/queries';
 import type { MetadataRoute } from 'next';
-import { allPosts } from 'contentlayer2/generated';
 
 export type Changefreq = 'always' | 'hourly' | 'daily' | 'weekly' | 'monthly' | 'yearly' | 'never';
 
-export default function sitemap(): MetadataRoute.Sitemap {
+export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
   const rootPath = env.NEXT_PUBLIC_APP_URL || `${siteConfig.url}`;
 
-  const posts = allPosts
-    .filter((post) => post.published)
-    .sort((a, b) => compareDesc(new Date(a.date), new Date(b.date)));
+  const posts = await getAllPosts();
 
   const postsRoutes = posts.map((post) => ({
-    url: `${rootPath}${post.slug}`,
+    url: `${rootPath}/posts/${post.slug}`,
     lastModified: new Date().toISOString(),
     changeFrequency: 'daily' as Changefreq,
     priority: 0.8,